[ros] omit final null character when sending strings, fixes #452
For compatibility with an ancient version of yarp, strings have been transmitted
with a null character included. Non-ancient versions of yarp will accept either
a string with or without a null termination. This commit stops transmitting that
null termination, but preserves the ability to receive a null terminated string
for compatibility between programs using different versions of yarp.